refactor: 💡 move utils to libapi
