+++ libs/libapi/pyproject.toml
@@ -13,6 +13,7 @@ orjson = "^3.8.6"
 pyjwt = { extras = ["crypto"], version = "^2.6.0" }
 python = "3.9.15"
 requests = "^2.28.2"
+starlette = "^0.28.0"
 starlette-prometheus = "^0.9.0"
 
 [tool.poetry.group.dev.dependencies]
