+++ services/api/pyproject.toml
@@ -12,8 +12,6 @@ libapi = {path = "../../libs/libapi", develop = true}
 pyarrow = "^11.0.0"
 python = "3.9.15"
 soundfile = ">=0.12.1"
-starlette = "^0.27.0"
-starlette-prometheus = "^0.9.0"
 uvicorn = "^0.20.0"
 watchdog = { extras = ["watchmedo"], version = "^2.2.1" }
