+++ services/api/src/api/routes/rows.py
@@ -18,7 +18,13 @@ from libapi.exceptions import (
     MissingRequiredParameterError,
     UnexpectedApiError,
 )
-from libapi.utils import Endpoint
+from libapi.utils import (
+    Endpoint,
+    are_valid_parameters,
+    get_json_api_error_response,
+    get_json_ok_response,
+    try_backfill_dataset,
+)
 from libcommon.parquet_utils import Indexer, StrPath
 from libcommon.processing_graph import ProcessingGraph
 from libcommon.prometheus import StepProfiler
@@ -30,13 +36,6 @@ from libcommon.viewer_utils.asset import (
 from libcommon.viewer_utils.features import get_cell_value
 from starlette.requests import Request
 from starlette.responses import Response
-
-from api.utils import (
-    are_valid_parameters,
-    get_json_api_error_response,
-    get_json_ok_response,
-    try_backfill_dataset,
-)
 
 logger = logging.getLogger(__name__)
