+++ services/api/src/api/routes/webhook.py
@@ -5,7 +5,7 @@ import logging
 from typing import Any, Literal, Optional, TypedDict
 
 from jsonschema import ValidationError, validate
-from libapi.utils import Endpoint
+from libapi.utils import Endpoint, get_response
 from libcommon.exceptions import CustomError, DatasetRevisionEmptyError
 from libcommon.operations import backfill_dataset, delete_dataset
 from libcommon.processing_graph import ProcessingGraph
@@ -13,8 +13,6 @@ from libcommon.prometheus import StepProfiler
 from libcommon.utils import Priority
 from starlette.requests import Request
 from starlette.responses import Response
-
-from api.utils import get_response
 
 schema = {
     "$schema": "https://json-schema.org/draft/2020-12/schema",
